Move integration test to RestoreIntegrationTest
